Organise ridley functions
