--- src/freeFunctions/game/monitorRidleyMovementSound.cpp
@@ -1,6 +1,0 @@
-#include "../../../include/models/iRidley.hpp"
-
-void monitorRidleyMovementSound(IRidley &ridley)
-{
-  ridley.stopMovementSoundIfAtSideOfWindow();
-}
