+++ src/freeFunctions/game/ridleyFunctions.cpp
@@ -1,5 +1,10 @@
 #include "../../../include/models/iRidley.hpp"
 
+void monitorRidleyMovementSound(IRidley &ridley)
+{
+  ridley.stopMovementSoundIfAtSideOfWindow();
+}
+
 void spawnRidley(IRidley &ridley)
 {
   ridley.spawn();
